telegram_bot.py
import os
import asyncio
from telegram import Update
from telegram.ext import ApplicationBuilder, CommandHandler, ContextTypes
from dotenv import load_dotenv
import json

load_dotenv()

class NeuroTelegramBot:
    def __init__(self):
        self.token = os.getenv("TELEGRAM_TOKEN", "")
        self.chat_id = os.getenv("TELEGRAM_CHAT_ID", "")
        
    async def start(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        await update.message.reply_text("🤖 NeuroTrader v3.1 Online\nCommands: /status, /report, /add symbol, /pause symbol")

    async def get_status(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        # In a real app, you'd read live state from a file or DB
        status_msg = "✅ System: Running\n📈 Equity: $10,450\n📉 Open Trades: 2\n🔥 CPU: 12%"
        await update.message.reply_text(status_msg)

    async def daily_report(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        report = "📅 Daily Report (2026-02-19)\n💰 PnL: +$245.50 (2.4%)\n✅ Wins: 12\n❌ Losses: 4"
        await update.message.reply_text(report)

    def run(self):
        if not self.token:
            print("Telegram token missing. Bot disabled.")
            return
            
        app = ApplicationBuilder().token(self.token).build()
        app.add_handler(CommandHandler("start", self.start))
        app.add_handler(CommandHandler("status", self.get_status))
        app.add_handler(CommandHandler("report", self.daily_report))
        
        print("Telegram bot starting...")
        app.run_polling()

if __name__ == "__main__":
    bot = NeuroTelegramBot()
    bot.run()
